Added a Greeting component
